refactor canvas_auth_spec.rb
